#39 finish implementation of terraform provider

+++ terraform-provider-terratowns/main.go
@@ -225,7 +225,7 @@ func resourceHouseRead(ctx context.Context, d *schema.ResourceData, m interface{
 		d.Set("name",responseData["name"].(string))
 		d.Set("description",responseData["description"].(string))
 		d.Set("domain_name",responseData["domain_name"].(string))
-		d.Set("content_version",responseData["content_version"].(int))
+		d.Set("content_version",responseData["content_version"].(float64))
 	} else if resp.StatusCode != http.StatusNotFound {
 		d.SetId("")
 	} else if resp.StatusCode != http.StatusOK {
@@ -249,7 +249,7 @@ func resourceHouseUpdate(ctx context.Context, d *schema.ResourceData, m interfac
 	payload := map[string]interface{}{
 		"name": d.Get("name").(string),
 		"description": d.Get("description").(string),
-		"content_version": d.Get("content_version").(int),
+		"content_version": d.Get("content_version").(float64),
 	}
 	payloadBytes, err := json.Marshal(payload)
 	if err != nil {
